requirements.txt
requests
python-bring-api
os
load_dotenv